Check if this is a db:verify packet BEFORE calling sx_nad_write()
sx_nad_write() eventually calls nad_free(), which means that the
call to NAD_ENAME(pkt->nad, 0) can reference uninitialized memory.
Before my nad caching change earlier today this was never a problem
because we never actually free'd nads--we just added them to the
cache.

The simple fix is just to move the check if it's a db:verify packet
to before the call to sx_nad_write()