Don't advertise that we support the digest-md5 sasl mechanism if our authreg doesn't support "get password," because it doesn't work. And don't advertise that we support the plain sasl mechanism if our authreg doesn't support either get password or check password.
This is good, for example, when you're using authreg_pipe to
authenticate against a database that only contains the sha1 checksum
of users' passwords, but not the actual passwords themselves.  In
this case you can't get the password for digest-md5, but you can
check the password for plain.

The old scod sasl code used to check this when it was in its own
directory, but it got lost along the way.

I was hoping to be able to do this without hardcoding "digest-md5"
and "plain," but I don't believe it's possible.  libgsasl doesn't
have a way to determine if a given mechanism will request
GSASL_PASSWORD using the application's gsasl callback.