Change the way that nads are constructed when responding to roster requests.  We ran into a case where a user had gigantic roster, and we noticed that using nad_insert_elem() is much less efficient than nad_append_elem() because it has to memmove everthing following the node that you're inserting, and with large rosters you could be memmove huge chunks of memory, and calling memmove with a bunch of times with a large chunk of memory seems to be very slow.
This probably changes the ordering that rosters are sent from the
server to clients... but from everything I can find the ordering
is not supposed to matter.  Plus I don't think we're storing things
in any particular order on the server side--it's just whatever order
the storage module happens to give to us.