Remove scod.  It was our own self-written sasl implementation.  It's generally a better idea to use the shared gsasl library (or cyrus-sasl, if it's available).  This code is currently a bit buggy.
Thomasz was thinking this might be useful if someone wants to run jabberd
on an embedded device or something.  So if you need that, feel free to
resurrect this and fix it up.

The email discussion about this was on July 7th and 8th on the
jabberd2@lists.xiaoka.com mailing list.