I removed NAD caching.  First a little background: NAD stands for not-a-DOM.  It's a custom structure used to represent an XML tree in memory.  The API is difficult to use, but it is probably faster than a full DOM implementation.
Each time a stanza is read from a socket a NAD structure is created.
When jabberd is finished with this NAD structure it calls nad_free().
If the nad_cache() has less than 100 NADS in it, and this NAD is
not gigantic, then it is added to the cache.  Otherwise it is freed.
Then, the next time someone calls nad_new(), a nad will be removed
from the cache instead of allocating a new one from memory.

This is an optimization done with the rationale that allocating
memory is slow, and that using a cache of NADs will be faster.

However, we have found that, while allocating memory may be
somewhat slow, it doesn't really make that much of a difference.
But more importantly we've found that the NAD cache can have a
SEVERE impatc on the amount of memory used by the server.

One characteristic of using the NAD cache is that, if the buffers
within a NAD are not big enough to hold the data being parsed,
the buffers will be realloc'ed to a larger size.  As NADs get
recylced through the nad cache they tend to grow in size.  There
were limits in place that were intended to restrict the maximum
size of the nad cache.

But for whatever reason those limits weren't working and sm
consumed more and more memory until it needed to be restarted.
At Meebo our jabberd server used to be fine, and we would only
have to restart once a month or so.  But recently something
subtle changed in our application (we're not sure what) that
causes this memory problem to be magnified.

We think probably somewhere in sm nads are being copied or keep
for later use, and that, because the nads in the cache are much
much larger than they should be, the memory being used is much
much more than it should be (e.g. even a small message of maybe
1KB could end up taking 300KB of memory because the buffers
in the nad had been realloced to a large size).

The memory usage itself isn't as much a problem as the memory
fragmentation.  When lots of spots of memory are allocated it
becomes more difficult for the memory manager to find an
available chunk to allocate, and so malloc can take significantly
longer.