Fixed a CRASHER.
In the rare cases such as:
<caps:c xmlns:caps='http://jabber.org/protocol/caps' xmlns='jabber:client' ver='1.1' ext='pmuc-v1' node='http://www.android.com/gtalk/client/caps'/>
where the element <c/> is in 'http://jabber.org/protocol/caps' namespace
but its attributes are in 'jabber:client' namespace (probably not what the stanza author meant),
the parser got the elem structure generated wrong, thus leading to a combination
of element with parent namespace but existing prefix.
This in turn led to incorrect XML generation by serializer _nad_lp0() with namespace prefix
but without the actual namespace definition.
This causes 'unbound prefix' parser error in companion processes leading to component disconnection and shutdown.