Merged patch to specify origin IP addres of S2S connections by Eygene Ryabinkin <rea-fbsd@codelabs.ru>
Element 'origin' of the 'local' block inside s2s.xml specifies the
source IP for the connections that are made by the s2s module.  Default
value is 'local.ip' -- it is natural for the hosts that have many alias
IPs on the single interface and the IP Jabber uses isn't the primary IP
of the server.  Since the world expects that server will originate its
connections from the server's own IP, this default behaviour is the best
one.