Compute string length before calling apr_base64_decode functions
The apr_base64_decode_len and apr_base64_decode_binary functions do not
compute the length of the input string but do rely on them to compute
the number of bytes to process. In this case, the helper function
b64_decode is calling the apr_base64* functions with length of -1 which
results in buffer length checks failing and string never being decoded.