Fix sigabrt due to race condition
It seems that under big workloads a race condition is being created
and this leads to sess->result being NULL when it should not.
If sess->result is NULL and then we call

	sx_nad_write(s, sess->result);

we ended up with a SIGABRT due to a failed assertion and jabberd crashes.
This aims to protect against it by checking sess->result before
calling sx_nad_write().

I see that this is already being checked in c2s_router_sx_callback(),
but not in _c2s_client_sx_callback().