Add python script to export data in xep0227 format
Dump status of jabberd2 instnce (users, their configuration, rosters etc.)
in a format that is supposed to be readable by any compliant xmpp
server.

At the time of sumbission, data is not importable into ejabberd
because of the bug https://github.com/processone/ejabberd/issues/2412
(there is a pull request addressing it). Also if external authetication
is used, ejabberd code containing commit de85c17 must be used.

There is a problem importing avatars into ejabberd, which is presumably(?)
a bug in ejabberd too.

Signed-off-by: Eugene Crosser <crosser@average.org>